requirements.txt
pytest==7.1.3
selenium==3.141.0
Appium-Python-Client==1.0.2
